+++ source/main.lua
@@ -5,7 +5,7 @@
 -- The check tables are used to check what happens each frame
 -- to the game and whether or not it should switch screens.
 --
--- reSnake - Copyright (C) 2022 - TNMM
+-- reSnake - Copyright (C) 2022-2023 - TNMM
 --
 -- This program is free software: you can redistribute it and/or modify
 -- it under the terms of the GNU General Public License as published by