+++ source/util.lua
@@ -1,7 +1,7 @@
 --- Utility file.
 -- Contains functions that are useful in the whole game.
 --
--- reSnake - Copyright (C) 2022 - TNMM
+-- reSnake - Copyright (C) 2022-2023 - TNMM
 --
 -- This program is free software: you can redistribute it and/or modify
 -- it under the terms of the GNU General Public License as published by